Assign values to the constants of the TYEventType enum.
This is a precaution to avoid accidental changes to the constant values
that could otherwise occur if the values were assigned automatically.
In that case, introducing a new constant between existing ones would
change the values of the constants following the new one. This would
cause problems to features that depend on the immutability of those
constants, e.g.: some feature that stores those values in a database.